pyproject.toml
[project]
name = "lms-summarizer"
version = "1.0.0"
description = "숭실대 LMS 자동 수강 처리 + 스크립트 추출 CLI"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
    "faster-whisper>=1.1.0",
    "pydantic>=2.11.0",
    "openai>=1.84.0",
    "google-genai>=1.0.0",
    "pyperclip>=1.9.0",
]

[project.scripts]
lms = "src.auto_watch.main:cli_entry"

[project.optional-dependencies]
dev = [
    "ruff>=0.11",
    "mypy>=1.15",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]

[tool.ruff.lint.per-file-ignores]
# JS evaluate 문자열 내부 긴 줄 허용
"src/auto_watch/browser.py" = ["E501"]
"src/auto_watch/courses.py" = ["E501"]
"src/auto_watch/player.py" = ["E501"]
"src/auto_watch/transcription.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "faster_whisper.*",
    "google.*",
    "pyperclip.*",
    "ctranslate2.*",
]
ignore_missing_imports = true
